requirements.txt
dataclasses>=0.6
typing-extensions>=4.0.0
google-generativeai>=0.3.0  # For Gemini API
requests>=2.26.0  # For API calls
python-dotenv>=0.19.0  # For environment variable management 